Refactored all metadata containers
