+++ src/neptune/metadata_containers/model.py
@@ -15,12 +15,23 @@
 #
 __all__ = ["Model"]
 
+import os
 from typing import (
     Iterable,
     Optional,
 )
 
-from neptune.exceptions import InactiveModelException
+from neptune.attributes import constants as attr_consts
+from neptune.common.exceptions import NeptuneException
+from neptune.envs import CONNECTION_MODE
+from neptune.exceptions import (
+    InactiveModelException,
+    NeedExistingModelForReadOnlyMode,
+    NeptuneMissingRequiredInitParameter,
+    NeptuneModelKeyAlreadyExistsError,
+    NeptuneObjectCreationConflict,
+)
+from neptune.internal.backends.api_model import ApiExperiment
 from neptune.internal.backends.nql import (
     NQLAggregator,
     NQLAttributeOperator,
@@ -28,10 +39,20 @@ from neptune.internal.backends.nql import (
     NQLQueryAggregate,
     NQLQueryAttribute,
 )
+from neptune.internal.backgroud_job_list import BackgroundJobList
 from neptune.internal.container_type import ContainerType
+from neptune.internal.id_formats import QualifiedName
+from neptune.internal.init.parameters import (
+    DEFAULT_FLUSH_PERIOD,
+    DEFAULT_NAME,
+    OFFLINE_PROJECT_QUALIFIED_NAME,
+)
 from neptune.internal.state import ContainerState
+from neptune.internal.utils import verify_type
+from neptune.internal.utils.ping_background_job import PingBackgroundJob
 from neptune.metadata_containers import MetadataContainer
 from neptune.metadata_containers.metadata_containers_table import Table
+from neptune.types.mode import Mode
 
 
 class Model(MetadataContainer):
@@ -44,6 +65,76 @@ class Model(MetadataContainer):
     """
 
     container_type = ContainerType.MODEL
+
+    def __init__(
+        self,
+        with_id: Optional[str] = None,
+        *,
+        name: Optional[str] = None,
+        key: Optional[str] = None,
+        project: Optional[str] = None,
+        api_token: Optional[str] = None,
+        mode: Optional[str] = None,
+        flush_period: float = DEFAULT_FLUSH_PERIOD,
+        proxies: Optional[dict] = None,
+    ):
+        verify_type("key", key, (str, type(None)))
+        verify_type("with_id", with_id, (str, type(None)))
+        verify_type("name", name, (str, type(None)))
+        verify_type("project", project, (str, type(None)))
+        verify_type("mode", mode, (str, type(None)))
+
+        self._key: Optional[str] = key
+        self._with_id: Optional[str] = with_id
+        self._name: str = DEFAULT_NAME if with_id is None and name is None else name
+
+        # make mode proper Enum instead of string
+        mode = Mode(mode or os.getenv(CONNECTION_MODE) or Mode.ASYNC.value)
+
+        if mode == Mode.OFFLINE:
+            raise NeptuneException("Model can't be initialized in OFFLINE mode")
+
+        if mode == Mode.DEBUG:
+            project = OFFLINE_PROJECT_QUALIFIED_NAME
+
+        super().__init__(project=project, api_token=api_token, mode=mode, flush_period=flush_period, proxies=proxies)
+
+    def _get_or_create_api_object(self) -> ApiExperiment:
+        project_workspace = self._project_api_object.workspace
+        project_name = self._project_api_object.name
+        project_qualified_name = f"{project_workspace}/{project_name}"
+
+        if self._with_id is not None:
+            # with_id (resume existing model) has priority over key (creating a new model)
+            #  additional creation parameters (e.g. name) are simply ignored in this scenario
+            model_id = QualifiedName(project_qualified_name + "/" + self._with_id)
+            return self._backend.get_metadata_container(
+                container_id=model_id, expected_container_type=self.container_type
+            )
+        elif self._key is not None:
+            if self._mode == Mode.READ_ONLY:
+                raise NeedExistingModelForReadOnlyMode()
+
+            try:
+                return self._backend.create_model(project_id=self._project_api_object.id, key=self._key)
+            except NeptuneObjectCreationConflict as e:
+                base_url = self._backend.get_display_address()
+                raise NeptuneModelKeyAlreadyExistsError(
+                    model_key=self._key,
+                    models_tab_url=f"{base_url}/{project_workspace}/{project_name}/models",
+                ) from e
+        else:
+            raise NeptuneMissingRequiredInitParameter(
+                parameter_name="key",
+                called_function="init_model",
+            )
+
+    def _prepare_background_jobs(self) -> BackgroundJobList:
+        return BackgroundJobList([PingBackgroundJob()])
+
+    def _write_initial_attributes(self):
+        if self._name is not None:
+            self[attr_consts.SYSTEM_NAME_ATTRIBUTE_PATH] = self._name
 
     def _raise_if_stopped(self):
         if self._state == ContainerState.STOPPED:
