+++ src/neptune/metadata_containers/model_version.py
@@ -15,15 +15,34 @@
 #
 __all__ = ["ModelVersion"]
 
+import os
+from typing import Optional
+
+from neptune.attributes import constants as attr_consts
 from neptune.attributes.constants import SYSTEM_STAGE_ATTRIBUTE_PATH
+from neptune.common.exceptions import NeptuneException
+from neptune.envs import CONNECTION_MODE
 from neptune.exceptions import (
     InactiveModelVersionException,
+    NeedExistingModelVersionForReadOnlyMode,
+    NeptuneMissingRequiredInitParameter,
     NeptuneOfflineModeChangeStageException,
 )
+from neptune.internal.backends.api_model import ApiExperiment
+from neptune.internal.backgroud_job_list import BackgroundJobList
 from neptune.internal.container_type import ContainerType
+from neptune.internal.id_formats import QualifiedName
+from neptune.internal.init.parameters import (
+    DEFAULT_FLUSH_PERIOD,
+    DEFAULT_NAME,
+    OFFLINE_PROJECT_QUALIFIED_NAME,
+)
 from neptune.internal.operation_processors.offline_operation_processor import OfflineOperationProcessor
 from neptune.internal.state import ContainerState
+from neptune.internal.utils import verify_type
+from neptune.internal.utils.ping_background_job import PingBackgroundJob
 from neptune.metadata_containers import MetadataContainer
+from neptune.types.mode import Mode
 from neptune.types.model_version_stage import ModelVersionStage
 
 
@@ -37,6 +56,72 @@ class ModelVersion(MetadataContainer):
     """
 
     container_type = ContainerType.MODEL_VERSION
+
+    def __init__(
+        self,
+        with_id: Optional[str] = None,
+        *,
+        name: Optional[str] = None,
+        model: Optional[str] = None,
+        project: Optional[str] = None,
+        api_token: Optional[str] = None,
+        mode: Optional[str] = None,
+        flush_period: float = DEFAULT_FLUSH_PERIOD,
+        proxies: Optional[dict] = None,
+    ):
+        verify_type("model", model, (str, type(None)))
+        verify_type("with_id", with_id, (str, type(None)))
+        verify_type("name", name, (str, type(None)))
+        verify_type("project", project, (str, type(None)))
+        verify_type("mode", mode, (str, type(None)))
+
+        self._model: Optional[str] = model
+        self._with_id: Optional[str] = with_id
+        self._name: str = DEFAULT_NAME if model is None and name is None else name
+
+        # make mode proper Enum instead of string
+        mode = Mode(mode or os.getenv(CONNECTION_MODE) or Mode.ASYNC.value)
+
+        if mode == Mode.OFFLINE:
+            raise NeptuneException("ModelVersion can't be initialized in OFFLINE mode")
+
+        if mode == Mode.DEBUG:
+            project = OFFLINE_PROJECT_QUALIFIED_NAME
+
+        super().__init__(project=project, api_token=api_token, mode=mode, flush_period=flush_period, proxies=proxies)
+
+    def _get_or_create_api_object(self) -> ApiExperiment:
+        project_workspace = self._project_api_object.workspace
+        project_name = self._project_api_object.name
+        project_qualified_name = f"{project_workspace}/{project_name}"
+
+        if self._with_id is not None:
+            # with_id (resume existing model_version) has priority over model (creating a new model_version)
+            version_id = QualifiedName(project_qualified_name + "/" + self._with_id)
+            return self._backend.get_metadata_container(
+                container_id=version_id, expected_container_type=ModelVersion.container_type
+            )
+        elif self._model is not None:
+            if self._mode == Mode.READ_ONLY:
+                raise NeedExistingModelVersionForReadOnlyMode()
+
+            model_id = QualifiedName(project_qualified_name + "/" + self._model)
+            api_model = self._backend.get_metadata_container(
+                container_id=model_id, expected_container_type=self.container_type
+            )
+            return self.backend.create_model_version(project_id=self._project_api_object.id, model_id=api_model.id)
+        else:
+            raise NeptuneMissingRequiredInitParameter(
+                parameter_name="model",
+                called_function="init_model_version",
+            )
+
+    def _prepare_background_jobs(self) -> BackgroundJobList:
+        return BackgroundJobList([PingBackgroundJob()])
+
+    def _write_initial_attributes(self):
+        if self._name is not None:
+            self[attr_consts.SYSTEM_NAME_ATTRIBUTE_PATH] = self._name
 
     def _raise_if_stopped(self):
         if self._state == ContainerState.STOPPED:
